Split person details screen into components
